Test add incomes back (#73)
* test: add valid income back

* test: add income invalid amount

* test: add income missing info

* Using try except else

* test: add expense invalid amount

* refactor: reusing method for checking expense

+++ backend/pluto/pluto/_internal/domain/model/income.py
@@ -12,13 +12,13 @@ class Income:
     id: str
     user_id: str
     src: str
-    amount: int
+    amount: float
     inc_date: str
 
     dict = asdict
 
     @staticmethod
-    def new(user_id: str, src: str, amount: int) -> Income:
+    def new(user_id: str, src: str, amount: float) -> Income:
         curr_date = datetime.today().strftime("%Y-%m-%d")
         return Income(new_id(), user_id, src, amount, curr_date)
